make editimage component
